requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
requests>=2.28.0
matplotlib>=3.5.0
numpy>=1.21.0
certifi>=2021.0.0
